automata v0.1.0 Automata

Types

agent_result()

Specs

agent_result() :: %{failures: non_neg_integer(), total: non_neg_integer()}

A map representing the results of running an agent

failed()

Specs

failed() :: [{Exception.kind(), reason :: term(), Exception.stacktrace()}]

The error state returned by Automata.WorldInfo and Automata.AutomatonInfo

state()

Specs

state() :: module()

All automata start with a %State{}.

Functions

begin()

Specs

begin() :: agent_result()

Begins update of Behavior Tree. More agent types to come. TODO: remove test env, get from config

configuration()

Specs

configuration() :: Keyword.t()

Returns Automata configuration.

configure(options)

Specs

configure(Keyword.t()) :: :ok

Configures Automata.

Options

Automata supports the following options:

:trace - sets Automata into trace mode, this allows agents to print info on an episode(s) while running. Any arbitrary configuration can also be passed to configure/1 or start/1, and these options can then be used in places such as the builtin automaton types configurations. These other options will be ignored by the Automata core itself.

run()

Specs

run() :: agent_result()

Runs the world. It is invoked automatically if Automata is started via start/1.

start(options \\ [])

Specs

start(Keyword.t()) :: :ok

status(automaton_name)