automata v0.1.0 Automaton.CompositeServer behaviour

When a child behavior is complete and returns its status code the Composite decides whether to continue through its children or whether to stop there and then and return a value.

The behavior tree represents all possible Actions that your AI can take. The route from the top level to each leaf represents one course of action, and the behavior tree algorithm traverses among those courses of action in a left-to-right manner. In other words, it performs a depth-first traversal.

Functions

types()

Callbacks

add_child(term)

add_child(term()) :: {:ok, list()} | {:error, String.t()}

clear_children()

clear_children() :: {:ok, list()} | {:error, String.t()}

continue_status()

continue_status() :: atom()

remove_child(term)

remove_child(term()) :: {:ok, list()} | {:error, String.t()}