automata v0.1.0 Automaton.Types.BT.Behavior behaviour

An Automaton.Types.BT.Behavior is an abstract interface for composites and components that can be activated, run, and deactivated. Actions(Execution Nodes) provide specific implementations of this interface. Branches in the tree can be thought of as high level behaviors, heirarchically combining smaller behaviors to provide more complex and interesting behaviors.

Note: there is a bunch of placeholder stuff in here right now and needs thought and cleanup. The specs for the callbacks are not accurate, just needed something that compiled and ran. Needs more thought and attention as the design process continues.

Callbacks

abort()

abort() :: atom()

aborted?()

aborted?() :: bool() | nil

on_init(term)

on_init(term()) :: term() | {:error, String.t()}

on_terminate(term)

on_terminate(term()) ::
  :bh_aborted | :bh_failure | :bh_fresh | :bh_running | :bh_success

reset()

reset() :: atom()

running?()

running?() :: bool() | nil

status()

status() :: atom()

terminated?()

terminated?() :: bool() | nil

update(any)

update(any()) :: {:ok, Module.t()}